+++ BlockBuddies Main/BlockBuddies/ScreenManager.hpp
@@ -27,7 +27,7 @@ public:
 
 	//allow the screen to shakeit
 
-	void shake(int);
+	void shake(float);
 private:
     // All the Screens are created beforehand so they don't have to be reinstantiated every single time
     std::map <Screens, Screen*> screens;
